fix: Resolve race condition on form validation
The form validation is intended to block having multiple drafts of a
form type, however a race condition exists where multiple draft forms
can be validated before either of them are persisted. This means that
both forms pass valiation and then are both stored, leaving two drafts.

Remove the manual validation as there is no suitable event that can be
used to avoid the race condition while using this approach.

Add an index to the MongoDB collection to ensure that each trainee can
only have a single draft form for a given type.
Add an exception handler to return a sensible HTTP response for this
scenario.

Remove the specified naming for FormRPartA and FormRPartB collections,
as the given naming does not match the default casing. Using the default
naming convention avoids having to specify the given name in multiple
places of the codebase.

TIS21-1489