chore: upgrade java to version 17
Gradles Toolchains functionality allows specifying a Java language
version and vendor to be used for building the project. Specifically it
separates the JDK used to build/run the project from the Java version
used for running Gradle itself.
Gradle will auto-provision a suitable JDK if one is not found on the
local machine, this will help simplify working on the project and help
ensure consist JDKs are used.

Since the `sourceCompatibility` does not differ from the toolchain
version, it is now redundant and should be removed.

TIS21-SHED