fix: create consistent form ids
The form ID number is a different format depending on whether the form
The form ID number is a different format depending on whether the form
was ever saved as a draft or not, direct submission results in a UUID
style ID e.g. c95f2312-a08c-46d4-8e0c-5c13215c96d0 but if the form was
saved as a draft before later submitting the ID will be in the
ObjectId style e.g. 63a3376e7199d510f8f56808

Use an event listener to populate the form ID with a UUID string
representation before converting it to a document, this avoids Mongo
from setting the ObjectId automatically.
A follow up submission will migrate the existing ObjectIds to the new
standardised UUID string format.

TIS21-4006