fix: incorrect timestamp
The timestamp is incorrect for several reasons

 1. The default timezone is used, which is UTC when deployed, but should
    use London time instead.
 2. The timestamp can be out by 12 hours when the time is between 12:01
    and 23:59

Add a timezone property to the application.properties, this should be
defaulted to Europe/London and passed to the template engine.

Update the COJ templates to use `HH:mm` instead of `hh:mm` for the time
format, eliminating the 12 hour error due to 12-hour vs 24-hour clock
formatting. The timezone should also be included so the timestamp is
always using London time.

TIS21-6106
TIS21-6121