chore: local application property defaults
Create an `application-local.yml` file with sensible defaults for
running this service locally.
The profile should be activated by setting
`SPRING_PROFILES_ACTIVE=local` in the env vars when running this
service. No other env vars need to be set if everything is running
directly on localhost, but `DB_HOST` and `LOCALSTACK_HOST` should still
be set when running inside a container.
`SIGNATURE_SECRET_KEY` can still be used to ensure services all match,
but a default value is provided.

There is still a pre-requisite that the queues and buckets listed in
`application-local.yml` are created in the localstack instance.

NO-TICKET