refactor: Form Rs to use auto index creation
Indexes can be created automatically using the `@Indexed` annotation,
which simplifies the index creation compared with the
MongoConfiguration's `indexOps.ensureIndex()` approach. The downside to
auto-creation is that the indexes will not be automatically added back
to recreated collections without restarting the application. However,
that limitation is common with the MongoConfiguration's current approach
anyway.

TIS21-6540