feat: filter admin LTFT count by DBCs
Filter the result of the admin LTFT count endpoint based on the DBCs
that the admin user is assigned to.
The DBCs are set as a Cognito group and added to the auth token as
`cognito:groups`.
This may include non-DBC groups at some point, but those should never
match a PM's DBC anyway.

TIS21-6599
TIS21-6938