feat(ltft): add missing fields to LTFT detail DTO
Refactor `LtftFormDto` and add the missing fields.
The detail DTO has been based on the database model, with some
additional fields taken from the API docs.

The current docs and the DTO are not an exact match, the DTO may need
further tweaks in some places but the API docs are also known to be out
of date.
Health-Education-England/tis-trainee-api#28 exists with some of the
changes, already. Other changes like the `formRef` field still need to
be added once PR-28 has been approved/merged.

The CCT date is calculated dynamically when converting from the database
entity, the value will always be null until the calculation is
implemented.

TIS21-6540