fix(ltft): do set null fields to default values
Several LTFT fields can have valid `null` values when working with a
DRAFT form, if that section has not been filled in.
Currently several fields are using primitive boolean and numerical
types, which are being defaulted to `false`/`0`.

Update the LTFT entity and model to use non-primative types which
can be defaulted to `null`.
Several of these fields should never be `null`, such as PM WTE which is
always expected to be populated. However, defaulting to erroneous values
causes more issues than missing data, as the source of the value becomes
unknown/lost.

TIS21-6540