feat(ltft): sorting and filtering admin endpoints
The sorting and filtering of admin endpoints is currently limited as the
internal field names are required.
As the internal field names are not transparent to the client, it should
be possible to sort and filter based on the DTO field names.

Update LtftService to translate the supported field names for both
sorting and filtering. The user filters should be added on top of the
filters for DBC and non-DRAFT LTFTs, but not replace them.

To enable sorting on `submissionDate`, which is a calculated field, it
should be made in to an actual field on the database model. The
`submitted` timestamp will be updated each time the LTFT is submitted.
Refactor AbstractAuditedForm to add the new field and remove the
existing `getSubmitted()` method.

Update `LtftAdminSummaryDto` to include the missing `formRef` field.

Update `LtftContent` to add indexes for the filterable fields.
`assignedAdmin.name` and `assignedAdmin.email` will not yet have indexes
due to the shared `Person` type. Indexes will need to be created
programmatically using `indexOps`.

TIS21-6934
TIS21-7098
TIS21-7099
TIS21-6943