feat(ltft): admin details PDF
Add an endpoint to allow TIS admins to download LTFT details as a PDF.
The API path is the same as getting the JSON detail, but the `Accept`
header should be set to `application/pdf`.

The PDF template is rough but data-wise complete. Further restructuring
of the HTML and CSS styling will be needed.

TIS21-6635
TIS21-7177