feat(ltft): restrict readonly properties on create
Several properties on the LTFT DTO should be read-only
 - `formRef`
 - `revision`
 - `assignedAdmin`
 - `status`
 - `created`
 - `lastModified`

Currently it is possible to create new LTFTs with user-provided values
in these fields. Any given values should be rejected/ignored and
server-generated values should be persisted and returned in the response
instead.

Update LtftFormDto to add validation to ensure that the values for these
properties is `null` during LTFT creation.
A `ReadOnly` JsonView should be used to make sure the deserialized DTO
never contains values for these fields. This will cause the validation
to always pass, but the dual-front approach ensures less chance of
misconfiguration allowing values through.

TIS21-7137
TIS21-7149