fix(ltft): ignore read-only fields on update
When a trainee updates an LTFT it is possible to provide values for the
following fields, which should be read only:
 - `formRef`
 - `revision`
 - `assignedAdmin`
 - `status`
 - `created`
 - `lastModified`

Update LtftFormDto and LtftResource to ensure that the above fields are
either validated or excluded from deserialization using JsonViews.

Update LtftService to update the existing form content instead of
persisting the user-provided data. The provided content will be
extracted and inserted in to the existing form, with the exception of
`assignedAdmin` which will be copied from the existing form.

TIS21-7137
TIS21-7150