fix(ltft): stop duplicating assignment history
When an admin is assigned multiple times it causes an audit history
entry each time, even though nothing has changed.
Update AbstractAuditedForm to check if the new assignment matches the
existing assignment and skip the update if it does.

The whole Person object should be checked to account for any of name,
email or role changing and needing to be updated. But if all entries
match then no status/history update should be done.

This check is already covered for status history due to restricted
transitions stopping the same state from being set multiple times.

TIS21-7082