fix(ltft): PDF end date always blank
As LTFT changes have no recorded end date, the end date field in the PDF
is always blank.
The end date for an LTFT change is implied to run until the new
programme end date, so the end date should use the CCT date as the
value during PDF generation.

This will mean duplicate values in "End Date" and "CCT Date" of the
Changes section, but being explicit about them being the same is
preferred rather than not showing the End Date property at all.

NO-TICKET