test: disable validation response integration tests
All tests reliant on the RestControllerAdvice to translate validation
exceptions in to Problem Details began failing for unknown reasons.
It appears that the exception handler is no longer loaded/registered
during CI/CD workflow runs.

This only occurs when running on GitHub Actions, it happens even when
running the workflow against old code that was previously successful.
So far it has not been possible to reproduce this issue locally.

Attempts to import or otherwise configure the exception handler have so
far failed.
The tests should be disabled and the deployed behaviour manually
verified to ensure that it is behaving properly in the real world.
Resolving the issue on the build server can follow.

NO-TICKET