fix: LTFT exception handling
The integration tests for LTFT validation started failing due to changes
on the build server. While the exact reason for the changed behaviour is
unknown, it has highlighted a conflict between
`RestResponeEntityExceptionHandler` and `ExceptionTranslator`.

Update the `ExceptionTranslator` to specify which rest controllers it
should be responsible for. The LTFT controllers will continue to use the
Problem Detail response with `application/problem+json` content type.

NO-TICKET