refactor: authentication/authorization
Auth currently relies on manual extraction of the claims from the JWT
token, with a check in the interceptor to validate the required roles.
That worked for the initial need of restricting LTFT admin
functionality, but its not a suitable approach now we need to add
further role-based restrictions.

The standard approach with Spring Securty is annotation based on the API
endpoints.
Refactor the security configuration to implement Spring Security,
removing role-based checks from the Admin interceptor. Instead a JWT
converter will extract the Cognito roles/groups and set them as granted
authorities.

Due to the addition of Spring Security, the auth approach for MockMvc
must be changed everywhere to use `jwt()` provider instead of setting an
Authorization header. This allows bypassing the cryptographic token
validity checks which would otherwise now be required.

To avoid any breaking changes to internal endpoints, authentication will
not be enforced beyond the existing role/identity requirements
(`anyRequest().permitAll()`).
This configuration will be changed at a later time to enforce
authenticated access for admin/trainee endpoints, with unauthenticated
(or secret-based) access to private internal endpoints.

TIS21-7996