fix: unsubmitted Form Rs not editable
When Admin's unsubmitted a form via the common-upload service flow, the
S3 version of the form was updated to the UNSUBMITTED state without the
database being updated.

When getting a list of forms for the trainee, the S3 forms are queried
and DRAFT forms from the database are added to the result.
However, when getting a particular form by ID both versions of the form
(S3 and DB) are retrieved and the modified timestamps are compared -
with the latest being returned. In the case of equal timestamps the
database value is used.

This would ordinarily be fine, but the common-upload service does not
update the modified timestamp when changing form state. So the
UNSUBMITTED form has a matching timestamp to the DBs SUBMITTED form and
the SUBMITTED form is always returned.

Previously this was covered up by the fact that the frontend checked the
summary list for an unsubmitted form, and then didn't check the state of
the returned form when getting the form contents by ID. The additional
check wasn't strictly needed, as the state still contained data for the
UNSUBMITTED cloud summary.
Due to frontend changes to enable better routing for forms, we have
routes like `/formr-parta/:id/edit` which, given the above, pulls the
SUBMITTED form from the database and then rejects it due to be
uneditable - as we no longer rely on validation from state values.

Modify the conditional selection of S3 or database forms to use the S3
version in the case of equal timestamps.
This will avoid this issue with UNSUBMITTED forms as well as any
potential issues from de-synced form deletions which are reflected in S3
but not the database.
As we complete the Form-R migration this edge cases will resolve
automatically as we will be 100% database driven.

TIS21-8123